1st new commit to 1.2.2cvs. Franz's text on path fixes. Updated translation ts/po/qm files. String fixes in Scripter. Fix for Scrapbook file names.
git-svn-id: svn://scribus.net/trunk/Scribus@1522 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/scribusdoc.h
@@ -62,8 +62,8 @@ struct Pti {
 		    int cshade2;
 		    int cstyle;
 		    int cab;
-		    int PtransX;
-			int PtransY;
+		    double PtransX;
+			double PtransY;
 			int cscale;
 			double PRot;
 		   };
